requirements.txt
pyyaml==6.0.1
pytest==7.2.0
pytelegrambotapi==4.15.4